@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/dropzone_file_upload.
  *
- * (c) 2019 The MetaModels team.
+ * (c) 2019 - 2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,7 +12,8 @@
  *
  * @package    MetaModels/attribute_file
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2019 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2019 - 2022 The MetaModels team.
  * @license    https://github.com/MetaModels/dropzone_file_upload/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -27,8 +28,6 @@ use MetaModels\AttributeFileBundle\Attribute\File;
 use MetaModels\DcGeneral\Events\MetaModel\BuildAttributeEvent;
 use MetaModels\ViewCombination\ViewCombination;
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
-use Symfony\Component\Security\Csrf\TokenStorage\TokenStorageInterface;
 
 /**
  * This event build the dropzone for upload draggable file field, in the frontend editing scope.
@@ -37,59 +36,41 @@ final class BuildDropzoneUploadListener
 {
     use RequestScopeDeterminatorAwareTrait;
 
+    public const SESSION_KEY_UPLOAD_PATH = 'dropzone_upload_path';
+
     /**
      * The view combinations.
      *
      * @var ViewCombination
      */
-    private $viewCombination;
+    private ViewCombination $viewCombination;
 
     /**
      * The request stack.
      *
      * @var RequestStack
      */
-    private $request;
-
-    /**
-     * The security csrf token manager.
-     *
-     * @var CsrfTokenManagerInterface
-     */
-    private $tokenManager;
-
-    /**
-     * The token name.
-     *
-     * @var string
-     */
-    private $tokenName;
+    private RequestStack $request;
 
     /**
      * The property information from the input screen.
      *
      * @var array
      */
-    private $information;
+    private array $information;
 
     /**
      * The constructor.
      *
-     * @param ViewCombination           $viewCombination The view combination.
-     * @param RequestStack              $request         The request.
-     * @param CsrfTokenManagerInterface $tokenManager    The security csrf token manager.
-     * @param string                    $tokenName       The token name.
+     * @param ViewCombination $viewCombination The view combination.
+     * @param RequestStack    $request         The request.
      */
     public function __construct(
         ViewCombination $viewCombination,
-        RequestStack $request,
-        CsrfTokenManagerInterface $tokenManager,
-        string $tokenName
+        RequestStack $request
     ) {
         $this->viewCombination = $viewCombination;
         $this->request         = $request;
-        $this->tokenManager    = $tokenManager;
-        $this->tokenName       = $tokenName;
     }
 
     /**
@@ -146,17 +127,31 @@ final class BuildDropzoneUploadListener
     private function getTempFolderPath(BuildAttributeEvent $event): string
     {
         $request = $this->request->getMasterRequest();
+        if (null === $request) {
+            throw new \LogicException('No request set on the stack');
+        }
 
         $modelId = 'create';
         if ($request->query->has('id')) {
             $modelId = ModelId::fromSerialized($request->query->get('id'))->getId();
         }
 
+        $session = $request->getSession();
+        // NOTE: this check can be removed when depending on symfony/http-foundation 5.0+
+        if (null === $session) {
+            throw new \LogicException('No session set on the request');
+        }
+
+        if (null === ($randomPath = $session->get(self::SESSION_KEY_UPLOAD_PATH))) {
+            $randomPath = \md5($session->getId());
+            $session->set(self::SESSION_KEY_UPLOAD_PATH, $randomPath);
+        }
+
         $pieces = [
             'system',
             'tmp',
             'dropzone',
-            $this->tokenManager->getToken($this->tokenName),
+            $randomPath,
             $event->getAttribute()->getMetaModel()->getTableName(),
             $modelId,
             $event->getAttribute()->getColName()
